Improve stream reliability by moving LSL handling into foreground service
Previously the LSL streams were handled by the HandLandmarkerHelper which
was problematic because this class is frequently closed and reinstantiated,
causing streams to be destroyed/recreated by every configuration change or
when the app is moved to the background, when the streams are recreated before
the old streams are garbage-collected there can be two instances of the same
stream in the network etc.

I moved the streams into an LSL service similar to SENDA. The service is started
when the camera fragment is created and stopped when the fragment is destroyed and
acquires a wakelock to prevent the device from going into standby. Differently from
SENDA, I use binding to pass hand landmark data to the service.

Although the service verifiably keeps running while the app is in the background,
the streams still stop being discoverable while the app is in the background.
Android probably shut off network access, even though I applied the proper foreground
service types. As the camera is only accessed while the app is in the foreground, and the
streams recover reliably, this is annoying but not catastrophic. I need to check if the
streams stay discoverable if they keep receiving data while in background. The camera
background service type seems to suggest we might keep running the landmarker even if in background.